rename: Make ScheduleModal directory
