+++ components/MonthCalendar/MonthCalendar.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 
 import styles from './MonthCalendar.module.scss';
 
-import CreateScheduleButton from '@components/CreateSchedule/CreateScheduleButton';
 import DayinMonth from '@components/MonthCalendar/DayinMonth';
+import CreateScheduleButton from '@components/ScheduleModal/CreateScheduleButton';
 import Sidebar from '@components/Sidebar/Sidebar';
 import { useSidebarContext } from '@contexts/SidebarContext';
 import { DAYS_ARR } from '@utils/formatDay';
